New while loop
I put all of the code that was being repeated in each function into a while loop in the main function to make things less complex and easier to handle.
As well, I put a pass statement in the livingRoom2 function because I am going to make it into a battle scenario.