Game changer
I made a very large number of changes.
First, I put the part that checks for the number of times a player has switched rooms into the move_to_location function and moved the bug checking print statements for the world inventory into the add_to_inventory function.
I modified the living room function so that the text saying that the key has successfully been stolen will only show if the conditions of the final battle are not met.
I made a lot of updates to the dining room file. For one, I made a separate dictionary for a second enemy (which is the monster hiding behind the new door that appears in the dining room) and then incorporated it into the battle function. While doing so, I also changed the name of the finalFight function to combat() and put if statements to decide what to do depending on which enemy is being faced.
When fighting the purple gremlin, the player will automatically heal fully upon a victory.
What's more, the dining room file will now have different texts displayed depending on how many times the player has been in the room. These texts are now different for the 1st, 2nd, and even 3rd time of entry.
I added conditions in the if statements in the dining room function so that the option of the new door will only be presented if the player has not already defeated the purple gremlin. I achieved this by created a separate element in the world dictionary called "vic" for victories. If the gremlin is defeated, it is added to the list "vic" and the option of the "new door" will only display if "purple gremlin" is not in the "vic" list.
Many of the if statements in the game loop have been changed to elifs so that there will be no overlap of conditions.
Before, there was also an issue where if the player grabbed the axe before the key, the program would not automatically proceed to the final battle with the monster even after the key had been added to world["inv"]. This problem no longer occurs since I added the additional conditions of 'if not {"key", "axe"}.issubset(world["inv"])' before the print statements saying that the player has successfully stolen the key in the living room function.