If the player and the enemy die.
I have edited the combat.py file in the function battle.py so that it handles the condition of when both the player and the monster have hp less than zero. Instead of only saying the player has lost, it will acknowledge that the monster has been defeated as well but that during the fight the player was also killed.
I made corresponding adjustments to the combat function in the text adventure file.